+++ resources/views/components/badge.blade.php
@@ -1,0 +1,1 @@
+<{{ $tagName() }} {{ $attributes->merge(['href' => $href])->class([$cssClass()]) }}>{{ $slot }}</{{ $tagName() }}>